refactor: move glucose tests into same module
